[Gepardec/mega#674] adapt test impl due to movement of endpoint
