+++ src/test/java/com/gepardec/mega/rest/EmployeeResourceTest.java
@@ -45,14 +45,9 @@ class EmployeeResourceTest {
     @Inject
     EmployeeMapper mapper;
 
-    @Inject
-    EmployeeResource employeeResource;
 
     @InjectMock
     EmployeeService employeeService;
-
-    @InjectMock
-    ZepService zepService;
 
 
     @InjectMock
@@ -199,85 +194,11 @@ class EmployeeResourceTest {
                 .then().statusCode(HttpStatus.SC_METHOD_NOT_ALLOWED);
     }
 
-    /*@Test
-    void testGetBillsForEmployeeByMonth_whenEmployeeHasBills_thenReturnBills(){
-        User userForRole = createUserForRole(Role.PROJECT_LEAD);
-        when(userContext.getUser()).thenReturn(userForRole);
-        final Employee userAsEmployee = mapper.mapToDomain(createEmployeeForUser(userForRole));
 
 
-        when(employeeService.getEmployee(userAsEmployee.getUserId()))
-                .thenReturn(userAsEmployee);
 
 
-        when(zepService.getBillsForEmployeeByMonth(userAsEmployee, YearMonth.of(2024, 4))
-                .thenReturn(
-                    getBillsForEmployee()
-                );
 
-        List<BillDto> actual = employeeResource.getBillsForEmployeeByMonth(userAsEmployee.getUserId());
-
-        assertThat(actual).isNotNull().size().isEqualTo(3);
-        assertThat(actual.get(0).getBillType()).isEqualTo("Lebensmittel");
-    }
-
-    @Test
-    void testGetBillsForEmployeeByMonth_whenEmployeeHasNoBills_thenReturnEmptyList(){
-        User userForRole = createUserForRole(Role.PROJECT_LEAD);
-        when(userContext.getUser()).thenReturn(userForRole);
-        final Employee userAsEmployee = mapper.mapToDomain(createEmployeeForUser(userForRole));
-
-
-        when(employeeService.getEmployee(userAsEmployee.getUserId()))
-                .thenReturn(userAsEmployee);
-
-
-        when(zepService.getBillsForEmployeeByMonth(userAsEmployee))
-                .thenReturn(
-                        List.of()
-                );
-
-        List<BillDto> actual = employeeResource.getBillsForEmployeeByMonth(userAsEmployee.getUserId());
-
-        assertThat(actual).isNotNull();
-        assertThat(actual).isEmpty();
-    }*/
-    private List<Bill> getBillsForEmployee() {
-        return List.of(
-                createBillForEmployee(LocalDate.of(2024, 4, 11),
-                        12.0,
-                        "Lebensmittel",
-                        PaymentMethodType.COMPANY,
-                        "3BankenIT - JBoss",
-                        null),
-                createBillForEmployee(LocalDate.of(2024, 4, 10),
-                        12.0,
-                        "Büromaterial",
-                        PaymentMethodType.PRIVATE,
-                        "3BankenIT - JBoss",
-                        null),
-                createBillForEmployee(LocalDate.of(2024, 4, 9),
-                        30.0,
-                        "Lebensmittel",
-                        PaymentMethodType.PRIVATE,
-                        "3BankenIT - JBoss",
-                        null)
-        );
-    }
-
-
-    private Bill createBillForEmployee(LocalDate billDate, double bruttoValue, String billType,
-                                       PaymentMethodType paymentMethodType, String projectName,
-                                       String attachmentBase64String){
-        return Bill.builder()
-                .billDate(billDate)
-                .bruttoValue(bruttoValue)
-                .billType(billType)
-                .paymentMethodType(paymentMethodType)
-                .projectName(projectName)
-                .attachmentBase64(attachmentBase64String)
-                .build();
-    }
     private EmployeeDto createEmployeeForUser(final User user) {
         return EmployeeDto.builder()
                 .email(user.getEmail())
