+++ src/test/java/com/gepardec/mega/rest/WorkerResourceTest.java
@@ -1,18 +1,19 @@
 package com.gepardec.mega.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.gepardec.mega.db.entity.common.PaymentMethodType;
 import com.gepardec.mega.db.entity.employee.EmployeeState;
-import com.gepardec.mega.domain.model.Employee;
-import com.gepardec.mega.domain.model.Role;
-import com.gepardec.mega.domain.model.User;
-import com.gepardec.mega.domain.model.UserContext;
+import com.gepardec.mega.domain.model.*;
 import com.gepardec.mega.domain.model.monthlyreport.JourneyWarning;
 import com.gepardec.mega.domain.model.monthlyreport.MonthlyReport;
+import com.gepardec.mega.rest.api.WorkerResource;
 import com.gepardec.mega.rest.mapper.EmployeeMapper;
+import com.gepardec.mega.rest.model.BillDto;
 import com.gepardec.mega.rest.model.MappedTimeWarningDTO;
 import com.gepardec.mega.rest.model.MonthlyReportDto;
 import com.gepardec.mega.service.api.EmployeeService;
 import com.gepardec.mega.service.api.MonthlyReportService;
+import com.gepardec.mega.zep.ZepService;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.mockito.InjectMock;
 import io.quarkus.test.security.TestSecurity;
@@ -25,6 +26,8 @@ import jakarta.inject.Inject;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
 
+import java.time.LocalDate;
+import java.time.YearMonth;
 import java.util.List;
 import java.util.Set;
 
@@ -47,10 +50,16 @@ public class WorkerResourceTest {
     EmployeeService employeeService;
 
     @InjectMock
+    ZepService zepService;
+
+    @InjectMock
     UserContext userContext;
 
     @InjectMock
     EmployeeMapper mapper;
+
+    @Inject
+    WorkerResource workerResource;
 
 
     @Test
@@ -275,6 +284,86 @@ public class WorkerResourceTest {
         assertThat(nonVacationDays).isEqualTo(actual.getNonPaidVacationDays());
     }
 
+
+    @Test
+    void getBillsForEmployeeByMonth_whenEmployeeHasBills_thenReturnBills(){
+        User userForRole = createUserForRole(Role.EMPLOYEE);
+        when(userContext.getUser()).thenReturn(userForRole);
+        final Employee userAsEmployee = createEmployeeForUser(userForRole);
+
+
+        when(employeeService.getEmployee(userAsEmployee.getUserId()))
+                .thenReturn(userAsEmployee);
+
+
+        when(zepService.getBillsForEmployeeByMonth(userAsEmployee, YearMonth.of(2024, 4)))
+                .thenReturn(
+                    getBillsForEmployee()
+                );
+
+        List<BillDto> actual = workerResource.getBillsForEmployeeByMonth(userAsEmployee.getUserId(), YearMonth.of(2024, 4));
+
+        assertThat(actual).isNotNull().size().isEqualTo(3);
+        assertThat(actual.get(0).getBillType()).isEqualTo("Lebensmittel");
+    }
+
+    @Test
+    void getBillsForEmployeeByMonth_whenEmployeeHasNoBills_thenReturnEmptyList(){
+        User userForRole = createUserForRole(Role.EMPLOYEE);
+        when(userContext.getUser()).thenReturn(userForRole);
+        final Employee userAsEmployee = createEmployeeForUser(userForRole);
+
+
+        when(employeeService.getEmployee(userAsEmployee.getUserId()))
+                .thenReturn(userAsEmployee);
+
+
+        when(zepService.getBillsForEmployeeByMonth(userAsEmployee, YearMonth.of(2024, 4)))
+                .thenReturn(
+                        List.of()
+                );
+
+        List<BillDto> actual = workerResource.getBillsForEmployeeByMonth(userAsEmployee.getUserId(), YearMonth.of(2024, 4));
+
+        assertThat(actual).isNotNull();
+        assertThat(actual).isEmpty();
+    }
+
+    private Bill createBillForEmployee(LocalDate billDate, double bruttoValue, String billType,
+                                       PaymentMethodType paymentMethodType, String projectName,
+                                       String attachmentBase64String){
+        return Bill.builder()
+                .billDate(billDate)
+                .bruttoValue(bruttoValue)
+                .billType(billType)
+                .paymentMethodType(paymentMethodType)
+                .projectName(projectName)
+                .attachmentBase64(attachmentBase64String)
+                .build();
+    }
+
+    private List<Bill> getBillsForEmployee() {
+        return List.of(
+                createBillForEmployee(LocalDate.of(2024, 4, 11),
+                        12.0,
+                        "Lebensmittel",
+                        PaymentMethodType.COMPANY,
+                        "3BankenIT - JBoss",
+                        null),
+                createBillForEmployee(LocalDate.of(2024, 4, 10),
+                        12.0,
+                        "Büromaterial",
+                        PaymentMethodType.PRIVATE,
+                        "3BankenIT - JBoss",
+                        null),
+                createBillForEmployee(LocalDate.of(2024, 4, 9),
+                        30.0,
+                        "Lebensmittel",
+                        PaymentMethodType.PRIVATE,
+                        "3BankenIT - JBoss",
+                        null)
+        );
+    }
     private Employee createEmployeeForUser(final User user) {
         return Employee.builder()
                 .email(user.getEmail())
